Refactor augment interface
