+++ tests/test_core.py
@@ -158,109 +158,3 @@ class TestRobustness(object):
         for p in pending:
             await p
 
-class TestAugment(object):
-    
-    @mark.asyncio
-    async def test_simple_augment(self, mock_client, event_loop):
-        events_received = 0
-        
-        def stub_generator():
-            nonlocal events_received
-            client = yield
-            print("Got client")
-
-            while True:
-                event = yield
-                print("Got event")
-                events_received += 1
-
-        processor.register_augment(mock_client, "test", stub_generator(), None)
-        mock_client.event(service="test")
-        mock_client.event(service="test")
-        mock_client.event(service="test")
-
-        await mock_client.consume_augments()
-        assert events_received == 3
-
-    @mark.asyncio
-    async def test_window_n_3(self, mock_client, event_loop):
-        def stub_generator():
-            client = yield
-            print("Got client")
-            acc = 0
-
-            while True:
-                for i in range(0, 3):
-                    event = yield
-                    print("Got event")
-                    acc += event["metric"]
-                client.event(service="accumulated", metric=acc)
-                acc = 0
-
-        processor.register_augment(mock_client, "test", stub_generator(), None)
-
-        for i in range(0, 10):
-            mock_client.event(service="test", metric=i*10)
-
-        await mock_client.consume_augments()
-
-        # 10 events pushed, every 3 extra event added (10/3 = 3)
-        assert len(mock_client.events) == 13
-
-        filtered = list(filter(lambda x: x["service"] == "accumulated",
-                               mock_client.events))
-
-        assert len(filtered) == 3
-        # 30, 120, 210
-        assert sum([event["metric"] for event in filtered]) == 360
-
-    @mark.asyncio
-    async def test_lagging_augment(self, mock_client, event_loop):
-        from time import time, sleep
-        
-        def stub_generator():
-            print("Getting client")
-            client = yield
-            print("Got client")
-
-            while True:
-                print("Waiting for event")
-                event = yield
-                print("Got event")
-                ts = time()
-                print("Doing Long sleep")
-                sleep(10)
-                te = time()
-                print("Done sleeping: {0}".format(te - ts))
-
-        processor.register_augment(mock_client, "test", stub_generator(), None)
-        ts = time()
-        mock_client.event(service="test")
-        print("Consuming augments")
-        (done, pending) = await mock_client.consume_augments(
-                timeout=0.1
-        )
-        te = time()
-        td = te - ts
-        assert len(mock_client.events) == 1
-        assert len(done) == 0
-        assert len(pending) == 1
-        assert td < 0.5
-
-
-    @mark.asyncio
-    async def test_no_premature_exec(self, mock_client, event_loop):
-        events_received = 0
-        def stub_generator():
-            nonlocal events_received
-            client = yield
-
-            while True:
-                event = yield
-
-                events_received += 1
-
-        processor.register_augment(mock_client, "test", stub_generator(), None)
-        mock_client.event(service="test", metric=1.0)
-
-        assert events_received == 0
